requirements.txt
torch>=2.2
transformers>=4.40
accelerate>=0.30
chronos-forecasting>=2.0
faiss-cpu>=1.8.0
numpy>=1.26
pandas>=2.2
pyarrow>=15.0
scikit-learn>=1.4
scipy>=1.12
PyWavelets>=1.6
matplotlib>=3.8
tqdm>=4.66
PyYAML>=6.0
paho-mqtt>=2.1
joblib>=1.4
